3-Implementation/Model/requirements.txt
# Core data science packages
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter and interactive development
jupyter==1.0.0
ipykernel==6.25.0
jupytext==1.16.0  # Untuk konversi md <-> ipynb

# Machine learning
scikit-learn==1.2.2
tensorflow==2.12.0  # Versi ini sudah mendukung GPU jika CUDA terinstall
tensorflow-gpu==2.12.0  # Eksplisit mengaktifkan dukungan GPU
cudatoolkit==11.8  # Untuk GPU acceleration (sesuaikan versi jika diperlukan)
cudnn==8.6.0  # Untuk GPU acceleration (sesuaikan versi jika diperlukan)
tensorflow-hub==0.14.0
tflite-runtime==2.12.0  # Untuk konversi model ke format mobile

# Computer vision
pillow==10.0.0
opencv-python==4.8.0.74
imgaug==0.4.0  # Untuk augmentasi data gambar

# NLP for article recommendation
nltk==3.8.1
wordcloud==1.9.2  # Visualisasi text dalam EDA
networkx==3.1  # Untuk visualisasi hubungan artikel
beautifulsoup4==4.12.2  # Untuk parsing artikel jika diperlukan

# Utilities
tqdm==4.65.0
pyyaml==6.0.1
joblib==1.3.2  # Untuk menyimpan model scikit-learn